refactor(characters): extract Characters.Effects
Second slice of #134. Active-concept derivation, effect aggregation,
and resolved_state form the foundation the other concerns build on
(preparation, progression, and rolls all start from resolved_state), so
they move out before the dependent concerns do. collect_active_concepts,
effects_from_decisions, and inventory_effects move along as private
helpers since Effects is their only consumer.
