+++ apps/ex_ttrpg_dev/lib/characters.ex
@@ -6,6 +6,7 @@ defmodule ExTTRPGDev.Characters do
   alias ExTTRPGDev.Characters.Advancement
   alias ExTTRPGDev.Characters.Character
   alias ExTTRPGDev.Characters.Decision
+  alias ExTTRPGDev.Characters.Effects
   alias ExTTRPGDev.Characters.InventoryItem
   alias ExTTRPGDev.Characters.Store
   alias ExTTRPGDev.RuleSystem.Evaluator
@@ -449,13 +450,7 @@ defmodule ExTTRPGDev.Characters do
   Walks the decisions tree starting from root decisions (scope: nil), adding each selected
   concept and recursing into any sub-choices that concept declares.
   """
-  def active_concepts(decisions, concept_metadata) do
-    decisions
-    |> Enum.filter(fn d -> d.scope == nil end)
-    |> Enum.reduce(MapSet.new(), fn %{choice: type, selection: id}, acc ->
-      collect_active_concepts({type, id}, decisions, concept_metadata, acc)
-    end)
-  end
+  defdelegate active_concepts(decisions, concept_metadata), to: Effects
 
   @doc """
   Returns the combined effects list for a character against a system.
@@ -463,20 +458,7 @@ defmodule ExTTRPGDev.Characters do
   Filters system-defined effects to only those whose source concept is active
   (per the character's decisions), then appends the character's own effects.
   """
-  def active_effects(%LoadedSystem{} = system, %Character{} = character) do
-    active = active_concepts(character.decisions, system.concept_metadata)
-    decision_effects = effects_from_decisions(character.decisions, system.concept_metadata)
-
-    system.effects
-    |> Enum.filter(fn
-      %Effect{source: {type, id}} -> MapSet.member?(active, {type, id})
-      %Effect{source: {type, id, _}} -> MapSet.member?(active, {type, id})
-      _ -> false
-    end)
-    |> Kernel.++(decision_effects)
-    |> Kernel.++(inventory_effects(system, character.inventory))
-    |> Kernel.++(character.effects)
-  end
+  defdelegate active_effects(system, character), to: Effects
 
   @doc """
   Computes the character's active effects and evaluates the full system DAG against them.
@@ -486,10 +468,7 @@ defmodule ExTTRPGDev.Characters do
   way to obtain a character's resolved state; callers that need only one element can
   discard the other.
   """
-  def resolved_state(%LoadedSystem{} = system, %Character{} = character) do
-    effects = active_effects(system, character)
-    {effects, Evaluator.evaluate!(system, character.generated_values, effects)}
-  end
+  defdelegate resolved_state(system, character), to: Effects
 
   @doc """
   Rolls for a concept using the roll definition attached to its type in the system config.
@@ -1102,43 +1081,6 @@ defmodule ExTTRPGDev.Characters do
     end
   end
 
-  defp effects_from_decisions(decisions, concept_metadata) do
-    Enum.flat_map(decisions, fn
-      %Decision{scope: {type, id}, choice: choice_id, selection: selected} ->
-        choice_def =
-          concept_metadata
-          |> Map.get({type, id}, %{})
-          |> Map.get("choices", %{})
-          |> Map.get(choice_id, %{})
-
-        case choice_def do
-          %{
-            "contributes_field" => field,
-            "contributes_value" => value,
-            "type" => target_type
-          } ->
-            [%Effect{source: {type, id}, target: {target_type, selected, field}, value: value}]
-
-          _ ->
-            []
-        end
-
-      _ ->
-        []
-    end)
-  end
-
-  defp inventory_effects(%LoadedSystem{} = system, inventory) do
-    Enum.flat_map(inventory, fn %InventoryItem{} = item ->
-      system.effects
-      |> Enum.filter(fn
-        %Effect{source: {type, id}} -> type == item.concept_type and id == item.concept_id
-        _ -> false
-      end)
-      |> Enum.map(fn %Effect{} = effect -> %{effect | item_fields: item.fields} end)
-    end)
-  end
-
   @doc """
   Returns the IDs of root (non-sub) concepts of `type_id` from `concept_metadata`.
 
@@ -1184,26 +1126,6 @@ defmodule ExTTRPGDev.Characters do
     |> Map.get("choices", %{})
     |> Enum.flat_map(fn {_, choice_def} ->
       if choice_def["type"] == type_id, do: choice_def["options"] || [], else: []
-    end)
-  end
-
-  defp collect_active_concepts({_type, _id} = key, decisions, concept_metadata, acc) do
-    acc = MapSet.put(acc, key)
-    choices = concept_metadata |> Map.get(key, %{}) |> Map.get("choices", %{})
-
-    Enum.reduce(choices, acc, fn {choice_id, choice_def}, acc ->
-      decision = Enum.find(decisions, &(&1.scope == key and &1.choice == choice_id))
-
-      if decision && choice_def["grants_to"] != "inventory" do
-        collect_active_concepts(
-          {choice_def["type"], decision.selection},
-          decisions,
-          concept_metadata,
-          acc
-        )
-      else
-        acc
-      end
     end)
   end
